scripts/bank-statement-importer/requirements.txt

pandas>=1.3.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
